pyproject.toml
[tool.poetry]
name = "nonebot-qq-bot"
version = "1.0.0"
description = "基于NoneBot2的QQ机器人"
authors = []

[tool.poetry.dependencies]
python = "^3.9.0"
nonebot2 = "^2.3.2"
nonebot-adapter-onebot = "^2.3.0"
nonebot-plugin-help = "^0.6.1"
nonebot-plugin-manager = "^0.2.0"
psutil = "^5.9.0"
pydantic = "^1.10.26"
fastapi = "^0.95.2"
starlette = "^0.27.0"
anyio = "^4.12.0"
idna = "^3.11"
typing-extensions = "^4.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
  "nonebot.adapters.onebot.v11"
]
drivers = [
  "nonebot.drivers.fastapi",
  "nonebot.drivers.websockets"
]